Add Jumping control
I added jumping into the player controller, with ground and platform detection. It took a bit to figure it out cause I had used it in 2d before, and had to remove all the 2d in the script for it to actually work. I also adjusted the gravity so it was less floaty.